+++ resources/components/laraknife/panels/text-area.blade.php
@@ -1,9 +1,31 @@
-@props(['title' => 'Edit', 'error' => '', 'class' => ''])
+@props(['title' => 'Edit', 'error' => '', 'class' => '', 'prev' => '', 'next' => '', 'up' => '', 'audio' => ''])
 <div id="main-content" class="container mt-5 lkn-frame background">
     <div id="text-header-frame" class="lkn-header-frame">
         <h1>{{ $title }}</h1>
     </div>
     <div id="text-frame-panel" class="lkn-text-frame">
-             {{ $slot }}
+        {{ $slot }}
     </div>
+    <div class="lkn-arrows">
+        @if ($prev !== '')
+            <div class="lkn-float-left">
+                <a href="{{ $prev }}"> <i class="bi bi-arrow-left-circle-fill"></i> {{ __('Back') }}</a>
+            </div>
+        @endif
+        @if ($next !== '')
+        <div class="lkn-float-right">
+            <a href="{{ $next }}">{{ __('Next') }} <i class="bi bi-arrow-right-circle-fill"></i></a>
+        </div>
+        @endif
+        @if ($up !== '')
+            <div class="lkn-center">
+                <a href="{{ $up }}"> <i class="bi bi-arrow-up-circle-fill"></i></a>
+            </div>
+        @endif
+    </div>
+    @if ($audio !== '')
+        <div>
+            <x-laraknife.forms.audio-raw fileLink="{{ $audio }}" class="lkn-expand100" />
+        </div>
+    @endif
 </div>